create a `liberapay.i18n` submodule
