+++ error.spt
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 from pando.http import status_strings
 
 from liberapay.exceptions import LazyResponse
-from liberapay.utils.i18n import HTTP_ERRORS
+from liberapay.i18n.base import HTTP_ERRORS
 
 [----------------------------------------]
