+++ liberapay/exceptions.py
@@ -24,7 +24,7 @@ class LazyResponse(Response):
     def render_in_english(self):
         f = self.lazy_body
         fake_state = {}
-        from liberapay.utils.i18n import LOCALE_EN, add_helpers_to_context
+        from liberapay.i18n.base import LOCALE_EN, add_helpers_to_context
         add_helpers_to_context(fake_state, LOCALE_EN)
         return f(*resolve_dependencies(f, fake_state).as_args)
