+++ tests/py/test_participant.py
@@ -19,9 +19,9 @@ from liberapay.exceptions import (
     UsernameIsEmpty,
     UsernameTooLong,
 )
+from liberapay.i18n.currencies import Money
 from liberapay.models.participant import NeedConfirmation, Participant
 from liberapay.testing import EUR, USD, Harness
-from liberapay.utils.currencies import Money
 
 
 class TestNeedConfirmation(Harness):
