+++ www/%username/charts.json.spt
@@ -10,8 +10,8 @@ this to mean, "no chart."
 
 """
 
+from liberapay.i18n.currencies import Money
 from liberapay.utils import get_participant
-from liberapay.utils.currencies import Money
 
 [---]
