+++ www/about/stats.spt
@@ -1,7 +1,7 @@
 # coding: utf8
 from __future__ import division, print_function, unicode_literals
 
-from liberapay.utils.currencies import Money, MoneyBasket
+from liberapay.i18n.currencies import Money, MoneyBasket
 
 db = website.db
 db_qc5 = website.db_qc5
