+++ www/for/%name/edit.spt
@@ -1,5 +1,5 @@
+from liberapay.i18n.base import LANGUAGES_2, get_lang_options
 from liberapay.utils import get_community
-from liberapay.utils.i18n import LANGUAGES_2, get_lang_options
 
 [---]
