+++ www/for/new.spt
@@ -1,6 +1,6 @@
 from liberapay.exceptions import AuthRequired
+from liberapay.i18n.base import LANGUAGES_2, get_lang_options
 from liberapay.models.community import name_maxlength, normalize
-from liberapay.utils.i18n import LANGUAGES_2, get_lang_options
 
 [---]
