+++ www/index.html.spt
@@ -2,8 +2,8 @@
 
 from math import ceil
 
+from liberapay.i18n.base import LANGUAGES_2
 from liberapay.models.participant import Participant
-from liberapay.utils.i18n import LANGUAGES_2
 
 query_cache = website.db_qc5
