+++ www/search.spt
@@ -1,5 +1,5 @@
+from liberapay.i18n.base import LANGUAGES_2, SEARCH_CONFS, strip_accents
 from liberapay.utils import markdown
-from liberapay.utils.i18n import LANGUAGES_2, SEARCH_CONFS, strip_accents
 
 [---]
